s/closure table/continuation table/g
